requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.6
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-dotenv==1.4.2
django-filter==24.2
django-import-export==4.0.7
django-oauth-toolkit==2.4.0
django-oauth2-provider==0.2.6.1
django-otp==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
markup==0.2.2
MarkupSafe==2.1.5
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.3
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
randomtimestamp==2.2
requests==2.32.3
shortuuid==1.0.13
simplejson==3.19.2
sqlparse==0.5.0
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
django-cors-headers==4.4.0
djangorestframework-jwt==1.11.0
